requirements.txt
firebase_admin==6.5.0
numpy==1.26.4
pandas==2.2.1
protobuf==5.26.0
python-dotenv==1.0.1
